Broke Format out into its own module
