+++ src/components/commentCount.tsx
@@ -6,7 +6,7 @@ import { textSans } from '@guardian/src-foundations/typography';
 import { border } from '@guardian/src-foundations/palette';
 import { remSpace } from '@guardian/src-foundations';
 
-import { Format } from 'item';
+import { Format } from 'format';
 import { getPillarStyles } from 'pillar';
