+++ src/components/follow.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { css, SerializedStyles } from '@emotion/core';
 import { textSans } from '@guardian/src-foundations/typography';
 
-import { Format } from 'item';
+import { Format } from 'format';
 import { getPillarStyles } from 'pillar';
 import { Contributor, isSingleContributor } from 'contributor';
 import { darkModeCss } from 'styles';
