+++ src/components/shared/keyline.tsx
@@ -3,7 +3,7 @@ import { css, SerializedStyles } from '@emotion/core'
 import { darkModeCss, wideContentWidth, wideColumnWidth } from 'styles';
 import { neutral } from '@guardian/src-foundations/palette';
 import { from } from '@guardian/src-foundations/mq';
-import { Design } from 'item';
+import { Design } from 'format';
 
 const BaseStyles = css`
     height: 10px;
